split doobie example to separate files;
