+++ modules/doobie/src/test/scala/tofu/doobie/example/model.scala
@@ -1,0 +1,4 @@
+package tofu.doobie.example
+
+final case class Person(id: Long, name: String, deptId: Long)
+final case class Dept(id: Long, name: String)
